folder structure refactoring
