--- src/components/ModeSelections/data.js
@@ -1,8 +1,0 @@
-const modesList = [
-    { id: 1, name: "Test Mode" },
-    { id: 2, name: "Start Mode" },
-  ];
-
-  const dates = {dateFrom:"2022-12-13", dateTo:"2022-12-13"}
-
-  export {modesList, dates}
